refactor: decompose page.go into cmap.go, encoder.go, layout.go, outline.go
page.go had 6 unrelated concerns in 1422 lines. Split into 4 focused files,
all remaining in package pdf (same package — no export changes required):

  outline.go  — Outline type, buildPageMap, buildOutline, resolveOutlineDest
  encoder.go  — TextEncoding interface, byteEncoder, multibyteCMapEncoder,
                 ucs2BEEncoder, nopEncoder, encoderForCMapName
  cmap.go     — Font, dictEncoder, cmap, cmapInterp, readCmap (tight cluster)
  layout.go   — Text, Rect, Point, Content, Column/Row, GetTextByColumn/Row,
                 TextVertical, TextHorizontal

page.go residual: Page type, matrix/gstate, three interpreter state machines
(plainTextState, walkState, contentState), walkTextBlocks, Content(), Reader API.
