+++ layout.go
@@ -1,0 +1,211 @@
+// Copyright 2014 The Go Authors.  All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+package pdf
+
+import (
+	"bytes"
+	"errors"
+	"fmt"
+	"sort"
+)
+
+// A Text represents a single piece of text drawn on a page.
+type Text struct {
+	Font     string  // the font used
+	FontSize float64 // the font size, in points (1/72 of an inch)
+	X        float64 // the X coordinate, in points, increasing left to right
+	Y        float64 // the Y coordinate, in points, increasing bottom to top
+	W        float64 // the width of the text, in points
+	S        string  // the actual UTF-8 text
+}
+
+// A Rect represents a rectangle.
+type Rect struct {
+	Min, Max Point
+}
+
+// A Point represents an X, Y pair.
+type Point struct {
+	X float64
+	Y float64
+}
+
+// Content describes the basic content on a page: the text and any drawn rectangles.
+type Content struct {
+	Text []Text
+	Rect []Rect
+}
+
+// Column represents the contents of a column
+type Column struct {
+	Position int64
+	Content  TextVertical
+}
+
+// Columns is a list of column
+type Columns []*Column
+
+// GetTextByColumn returns the page's all text grouped by column
+func (p Page) GetTextByColumn() (Columns, error) {
+	result := Columns{}
+	var err error
+
+	defer func() {
+		if r := recover(); r != nil {
+			result = Columns{}
+			err = errors.New(fmt.Sprint(r))
+		}
+	}()
+
+	showText := func(enc TextEncoding, currentX, currentY float64, s string) {
+		var textBuilder bytes.Buffer
+
+		for _, ch := range enc.Decode(s) {
+			_, err := textBuilder.WriteRune(ch)
+			if err != nil {
+				panic(err)
+			}
+		}
+		text := Text{
+			S: textBuilder.String(),
+			X: currentX,
+			Y: currentY,
+		}
+
+		var currentColumn *Column
+		columnFound := false
+		for _, column := range result {
+			if int64(currentX) == column.Position {
+				currentColumn = column
+				columnFound = true
+				break
+			}
+		}
+
+		if !columnFound {
+			currentColumn = &Column{
+				Position: int64(currentX),
+				Content:  TextVertical{},
+			}
+			result = append(result, currentColumn)
+		}
+
+		currentColumn.Content = append(currentColumn.Content, text)
+	}
+
+	p.walkTextBlocks(showText)
+
+	for _, column := range result {
+		sort.Stable(column.Content)
+	}
+
+	sort.Slice(result, func(i, j int) bool {
+		return result[i].Position < result[j].Position
+	})
+
+	return result, err
+}
+
+// Row represents the contents of a row
+type Row struct {
+	Position int64
+	Content  TextHorizontal
+}
+
+// Rows is a list of rows
+type Rows []*Row
+
+// GetTextByRow returns the page's all text grouped by rows
+func (p Page) GetTextByRow() (Rows, error) {
+	result := Rows{}
+	var err error
+
+	defer func() {
+		if r := recover(); r != nil {
+			result = Rows{}
+			err = errors.New(fmt.Sprint(r))
+		}
+	}()
+
+	showText := func(enc TextEncoding, currentX, currentY float64, s string) {
+		var textBuilder bytes.Buffer
+		for _, ch := range enc.Decode(s) {
+			_, err := textBuilder.WriteRune(ch)
+			if err != nil {
+				panic(err)
+			}
+		}
+
+		// if DebugOn {
+		// 	fmt.Println(textBuilder.String())
+		// }
+
+		text := Text{
+			S: textBuilder.String(),
+			X: currentX,
+			Y: currentY,
+		}
+
+		var currentRow *Row
+		rowFound := false
+		for _, row := range result {
+			if int64(currentY) == row.Position {
+				currentRow = row
+				rowFound = true
+				break
+			}
+		}
+
+		if !rowFound {
+			currentRow = &Row{
+				Position: int64(currentY),
+				Content:  TextHorizontal{},
+			}
+			result = append(result, currentRow)
+		}
+
+		currentRow.Content = append(currentRow.Content, text)
+	}
+
+	p.walkTextBlocks(showText)
+
+	for _, row := range result {
+		sort.Stable(row.Content)
+	}
+
+	sort.Slice(result, func(i, j int) bool {
+		return result[i].Position > result[j].Position
+	})
+
+	return result, err
+}
+
+// TextVertical implements sort.Interface for sorting
+// a slice of Text values in vertical order, top to bottom,
+// and then left to right within a line.
+type TextVertical []Text
+
+func (x TextVertical) Len() int      { return len(x) }
+func (x TextVertical) Swap(i, j int) { x[i], x[j] = x[j], x[i] }
+func (x TextVertical) Less(i, j int) bool {
+	if x[i].Y != x[j].Y {
+		return x[i].Y > x[j].Y
+	}
+	return x[i].X < x[j].X
+}
+
+// TextHorizontal implements sort.Interface for sorting
+// a slice of Text values in horizontal order, left to right,
+// and then top to bottom within a column.
+type TextHorizontal []Text
+
+func (x TextHorizontal) Len() int      { return len(x) }
+func (x TextHorizontal) Swap(i, j int) { x[i], x[j] = x[j], x[i] }
+func (x TextHorizontal) Less(i, j int) bool {
+	if x[i].X != x[j].X {
+		return x[i].X < x[j].X
+	}
+	return x[i].Y > x[j].Y
+}
