+++ page.go
@@ -10,14 +10,7 @@ import (
 	"fmt"
 	"io"
 	"iter"
-	"sort"
 	"strings"
-
-	"golang.org/x/text/encoding"
-	"golang.org/x/text/encoding/japanese"
-	"golang.org/x/text/encoding/korean"
-	"golang.org/x/text/encoding/simplifiedchinese"
-	"golang.org/x/text/encoding/traditionalchinese"
 )
 
 // A Page represent a single page in a PDF file.
@@ -197,414 +190,6 @@ func (p Page) Font(name string) Font {
 	return Font{p.Resources().Key("Font").Key(name), nil}
 }
 
-// A Font represent a font in a PDF file.
-// The methods interpret a Font dictionary stored in V.
-type Font struct {
-	V   Value
-	enc TextEncoding
-}
-
-// BaseFont returns the font's name (BaseFont property).
-func (f Font) BaseFont() string {
-	return f.V.Key("BaseFont").Name()
-}
-
-// FirstChar returns the code point of the first character in the font.
-func (f Font) FirstChar() int {
-	return int(f.V.Key("FirstChar").Int64())
-}
-
-// LastChar returns the code point of the last character in the font.
-func (f Font) LastChar() int {
-	return int(f.V.Key("LastChar").Int64())
-}
-
-// Widths returns the widths of the glyphs in the font.
-// In a well-formed PDF, len(f.Widths()) == f.LastChar()+1 - f.FirstChar().
-func (f Font) Widths() []float64 {
-	x := f.V.Key("Widths")
-	var out []float64
-	for i := 0; i < x.Len(); i++ {
-		out = append(out, x.Index(i).Float64())
-	}
-	return out
-}
-
-// Width returns the width of the given code point.
-func (f Font) Width(code int) float64 {
-	first := f.FirstChar()
-	last := f.LastChar()
-	if code < first || last < code {
-		return 0
-	}
-	return f.V.Key("Widths").Index(code - first).Float64()
-}
-
-// Encoder returns the encoding between font code point sequences and UTF-8.
-func (f Font) Encoder() TextEncoding {
-	if f.enc == nil { // caching the Encoder so we don't have to continually parse charmap
-		f.enc = f.getEncoder()
-	}
-	return f.enc
-}
-
-// encoderForCMapName returns the TextEncoding for a named PDF CMap/Encoding.
-func encoderForCMapName(n string) TextEncoding {
-	switch n {
-	case "WinAnsiEncoding":
-		return &byteEncoder{&winAnsiEncoding}
-	case "MacRomanEncoding":
-		return &byteEncoder{&macRomanEncoding}
-	case "Identity-H":
-		return &byteEncoder{&pdfDocEncoding}
-	case "90ms-RKSJ-H", "90ms-RKSJ-V", "90pv-RKSJ-H":
-		return &multibyteCMapEncoder{japanese.ShiftJIS}
-	case "UniGB-UCS2-H", "UniGB-UCS2-V",
-		"UniCNS-UCS2-H", "UniCNS-UCS2-V",
-		"UniJIS-UCS2-H", "UniJIS-UCS2-V",
-		"UniKS-UCS2-H", "UniKS-UCS2-V":
-		return &ucs2BEEncoder{}
-	case "GB-EUC-H", "GB-EUC-V",
-		"GBKp-EUC-H", "GBKp-EUC-V",
-		"GBK-EUC-H", "GBK-EUC-V":
-		return &multibyteCMapEncoder{simplifiedchinese.GBK}
-	case "ETen-B5-H", "ETen-B5-V",
-		"ETenms-B5-H", "ETenms-B5-V":
-		return &multibyteCMapEncoder{traditionalchinese.Big5}
-	case "KSCms-UHC-H", "KSCms-UHC-V",
-		"KSC-EUC-H", "KSC-EUC-V",
-		"KSCms-UHC-HW-H", "KSCms-UHC-HW-V":
-		return &multibyteCMapEncoder{korean.EUCKR}
-	default:
-		if DebugOn {
-			println("unknown encoding", n)
-		}
-		return &byteEncoder{&pdfDocEncoding}
-	}
-}
-
-func (f Font) getEncoder() TextEncoding {
-	toUnicode := f.V.Key("ToUnicode")
-	if toUnicode.Kind() == Stream {
-		if m := readCmap(toUnicode); m != nil {
-			return m
-		}
-		if DebugOn {
-			println("ToUnicode stream failed to parse, falling back to Encoding")
-		}
-	}
-	enc := f.V.Key("Encoding")
-	switch enc.Kind() {
-	case Name:
-		return encoderForCMapName(enc.Name())
-	case Dict:
-		return newDictEncoder(enc)
-	case Null:
-		return &byteEncoder{&pdfDocEncoding}
-	default:
-		if DebugOn {
-			println("unexpected encoding", enc.String())
-		}
-		return &byteEncoder{&pdfDocEncoding}
-	}
-}
-
-// dictEncoder handles fonts with Encoding dictionaries containing
-// BaseEncoding and/or Differences arrays per PDF spec section 9.6.6.
-type dictEncoder struct {
-	table [256]rune
-}
-
-func newDictEncoder(enc Value) *dictEncoder {
-	e := &dictEncoder{}
-	baseEnc := enc.Key("BaseEncoding")
-	var baseTable *[256]rune
-	switch baseEnc.Name() {
-	case "WinAnsiEncoding":
-		baseTable = &winAnsiEncoding
-	case "MacRomanEncoding":
-		baseTable = &macRomanEncoding
-	default:
-		baseTable = &pdfDocEncoding
-	}
-	copy(e.table[:], baseTable[:])
-
-	diff := enc.Key("Differences")
-	if diff.Kind() == Array {
-		code := -1
-		for j := 0; j < diff.Len(); j++ {
-			x := diff.Index(j)
-			if x.Kind() == Integer {
-				code = int(x.Int64())
-				continue
-			}
-			if x.Kind() == Name && code >= 0 && code < 256 {
-				if r := nameToRune[x.Name()]; r != 0 {
-					e.table[code] = r
-				}
-				code++
-			}
-		}
-	}
-	return e
-}
-
-func (e *dictEncoder) Decode(raw string) (text string) {
-	r := make([]rune, 0, len(raw))
-	for i := 0; i < len(raw); i++ {
-		r = append(r, e.table[raw[i]])
-	}
-	return string(r)
-}
-
-// A TextEncoding represents a mapping between
-// font code points and UTF-8 text.
-type TextEncoding interface {
-	// Decode returns the UTF-8 text corresponding to
-	// the sequence of code points in raw.
-	Decode(raw string) (text string)
-}
-
-type nopEncoder struct {
-}
-
-func (e *nopEncoder) Decode(raw string) (text string) {
-	return raw
-}
-
-// multibyteCMapEncoder decodes PDF content-stream bytes using an x/text Encoding.
-// Used for predefined CMaps whose raw bytes are a well-known legacy encoding
-// (e.g. Shift-JIS for 90ms-RKSJ-H). Silently falls back to raw bytes on error.
-type multibyteCMapEncoder struct {
-	enc encoding.Encoding
-}
-
-func (e *multibyteCMapEncoder) Decode(raw string) (text string) {
-	decoded, err := e.enc.NewDecoder().Bytes([]byte(raw))
-	if err != nil {
-		return raw
-	}
-	return string(decoded)
-}
-
-// ucs2BEEncoder decodes PDF content-stream bytes encoded as UCS-2 big-endian.
-// Used for predefined CMaps such as UniGB-UCS2-H/V, UniCNS-UCS2-H/V,
-// UniJIS-UCS2-H/V, and UniKS-UCS2-H/V. Each glyph selector is a 2-byte
-// big-endian Unicode code point (e.g. 中 = 0x4E2D). No external dependency needed.
-// UCS-2/BMP-only — no surrogate-pair handling; correct for Uni*-UCS2-* CMaps.
-type ucs2BEEncoder struct{}
-
-func (e *ucs2BEEncoder) Decode(raw string) (text string) {
-	r := make([]rune, 0, len(raw)/2)
-	for i := 0; i+1 < len(raw); i += 2 {
-		r = append(r, rune(uint16(raw[i])<<8|uint16(raw[i+1])))
-	}
-	return string(r)
-}
-
-type byteEncoder struct {
-	table *[256]rune
-}
-
-func (e *byteEncoder) Decode(raw string) (text string) {
-	r := make([]rune, 0, len(raw))
-	for i := 0; i < len(raw); i++ {
-		r = append(r, e.table[raw[i]])
-	}
-	return string(r)
-}
-
-type byteRange struct {
-	low  string
-	high string
-}
-
-type bfchar struct {
-	orig string
-	repl string
-}
-
-type bfrange struct {
-	lo  string
-	hi  string
-	dst Value
-}
-
-type cmap struct {
-	space   [4][]byteRange // codespace range
-	bfrange []bfrange
-	bfchar  []bfchar
-}
-
-// lookupBfchar searches m.bfchar for an entry of length n matching text.
-func (m *cmap) lookupBfchar(text string, n int) ([]rune, bool) {
-	for _, bfchar := range m.bfchar {
-		if len(bfchar.orig) == n && bfchar.orig == text {
-			return []rune(utf16Decode(bfchar.repl)), true
-		}
-	}
-	return nil, false
-}
-
-// lookupBfrange searches m.bfrange for an entry of length n whose range contains text.
-func (m *cmap) lookupBfrange(text string, n int) ([]rune, bool) {
-	for _, bfrange := range m.bfrange {
-		if len(bfrange.lo) == n && bfrange.lo <= text && text <= bfrange.hi {
-			if bfrange.dst.Kind() == String {
-				s := bfrange.dst.RawString()
-				if bfrange.lo != text {
-					b := []byte(s)
-					b[len(b)-1] += text[len(text)-1] - bfrange.lo[len(bfrange.lo)-1]
-					s = string(b)
-				}
-				return []rune(utf16Decode(s)), true
-			}
-			if bfrange.dst.Kind() == Array {
-				idx := text[len(text)-1] - bfrange.lo[len(bfrange.lo)-1]
-				v := bfrange.dst.Index(int(idx))
-				if v.Kind() == String {
-					return []rune(utf16Decode(v.RawString())), true
-				}
-				if DebugOn {
-					fmt.Printf("array %v\n", bfrange.dst)
-				}
-			} else {
-				if DebugOn {
-					fmt.Printf("unknown dst %v\n", bfrange.dst)
-				}
-			}
-			return []rune{noRune}, true
-		}
-	}
-	return nil, false
-}
-
-func (m *cmap) Decode(raw string) (text string) {
-	var r []rune
-Parse:
-	for len(raw) > 0 {
-		for n := 1; n <= 4 && n <= len(raw); n++ {
-			for _, space := range m.space[n-1] {
-				if space.low <= raw[:n] && raw[:n] <= space.high {
-					text := raw[:n]
-					raw = raw[n:]
-					if runes, ok := m.lookupBfchar(text, n); ok {
-						r = append(r, runes...)
-						continue Parse
-					}
-					if runes, ok := m.lookupBfrange(text, n); ok {
-						r = append(r, runes...)
-						continue Parse
-					}
-					r = append(r, noRune)
-					continue Parse
-				}
-			}
-		}
-		if DebugOn {
-			println("no code space found")
-		}
-		r = append(r, noRune)
-		raw = raw[1:]
-	}
-	return string(r)
-}
-
-// cmapInterp holds mutable state for the readCmap Interpret callback.
-type cmapInterp struct {
-	n  int
-	m  cmap
-	ok bool
-}
-
-func (s *cmapInterp) handleEndCodespace(stk *Stack) {
-	if s.n < 0 {
-		if DebugOn {
-			println("missing begincodespacerange")
-		}
-		s.ok = false
-		return
-	}
-	for i := 0; i < s.n; i++ {
-		hi, lo := stk.Pop().RawString(), stk.Pop().RawString()
-		if len(lo) == 0 || len(lo) != len(hi) {
-			if DebugOn {
-				println("bad codespace range")
-			}
-			s.ok = false
-			return
-		}
-		s.m.space[len(lo)-1] = append(s.m.space[len(lo)-1], byteRange{lo, hi})
-	}
-	s.n = -1
-}
-
-func (s *cmapInterp) handleEndBfchar(stk *Stack) {
-	if s.n < 0 {
-		panic("missing beginbfchar")
-	}
-	for i := 0; i < s.n; i++ {
-		repl, orig := stk.Pop().RawString(), stk.Pop().RawString()
-		s.m.bfchar = append(s.m.bfchar, bfchar{orig, repl})
-	}
-}
-
-func (s *cmapInterp) handleEndBfrange(stk *Stack) {
-	if s.n < 0 {
-		panic("missing beginbfrange")
-	}
-	for i := 0; i < s.n; i++ {
-		dst, srcHi, srcLo := stk.Pop(), stk.Pop().RawString(), stk.Pop().RawString()
-		s.m.bfrange = append(s.m.bfrange, bfrange{srcLo, srcHi, dst})
-	}
-}
-
-func (s *cmapInterp) interpretCmap(stk *Stack, op string) {
-	if !s.ok {
-		return
-	}
-	switch op {
-	case "findresource":
-		stk.Pop()
-		stk.Pop()
-		stk.Push(newDict())
-	case "begincmap":
-		stk.Push(newDict())
-	case "endcmap":
-		stk.Pop()
-	case "begincodespacerange":
-		s.n = int(stk.Pop().Int64())
-	case "endcodespacerange":
-		s.handleEndCodespace(stk)
-	case "beginbfchar":
-		s.n = int(stk.Pop().Int64())
-	case "endbfchar":
-		s.handleEndBfchar(stk)
-	case "beginbfrange":
-		s.n = int(stk.Pop().Int64())
-	case "endbfrange":
-		s.handleEndBfrange(stk)
-	case "defineresource":
-		stk.Pop().Name()
-		value := stk.Pop()
-		stk.Pop().Name()
-		stk.Push(value)
-	default:
-		if DebugOn {
-			println("interp\t", op)
-		}
-	}
-}
-
-func readCmap(toUnicode Value) *cmap {
-	s := &cmapInterp{n: -1, ok: true}
-	Interpret(toUnicode, s.interpretCmap)
-	if !s.ok {
-		return nil
-	}
-	return &s.m
-}
-
 type matrix [3][3]float64
 
 var ident = matrix{{1, 0, 0}, {0, 1, 0}, {0, 0, 1}}
@@ -619,33 +204,6 @@ func (x matrix) mul(y matrix) matrix {
 		}
 	}
 	return z
-}
-
-// A Text represents a single piece of text drawn on a page.
-type Text struct {
-	Font     string  // the font used
-	FontSize float64 // the font size, in points (1/72 of an inch)
-	X        float64 // the X coordinate, in points, increasing left to right
-	Y        float64 // the Y coordinate, in points, increasing bottom to top
-	W        float64 // the width of the text, in points
-	S        string  // the actual UTF-8 text
-}
-
-// A Rect represents a rectangle.
-type Rect struct {
-	Min, Max Point
-}
-
-// A Point represents an X, Y pair.
-type Point struct {
-	X float64
-	Y float64
-}
-
-// Content describes the basic content on a page: the text and any drawn rectangles.
-type Content struct {
-	Text []Text
-	Rect []Rect
 }
 
 type gstate struct {
@@ -781,150 +339,6 @@ func (p Page) GetPlainText(fonts map[string]*Font) (result string, err error) {
 	s := &plainTextState{enc: &nopEncoder{}, fonts: fonts, resources: p.Resources()}
 	Interpret(p.V.Key("Contents"), s.interpretPlain)
 	return s.buf.String(), nil
-}
-
-// Column represents the contents of a column
-type Column struct {
-	Position int64
-	Content  TextVertical
-}
-
-// Columns is a list of column
-type Columns []*Column
-
-// GetTextByColumn returns the page's all text grouped by column
-func (p Page) GetTextByColumn() (Columns, error) {
-	result := Columns{}
-	var err error
-
-	defer func() {
-		if r := recover(); r != nil {
-			result = Columns{}
-			err = errors.New(fmt.Sprint(r))
-		}
-	}()
-
-	showText := func(enc TextEncoding, currentX, currentY float64, s string) {
-		var textBuilder bytes.Buffer
-
-		for _, ch := range enc.Decode(s) {
-			_, err := textBuilder.WriteRune(ch)
-			if err != nil {
-				panic(err)
-			}
-		}
-		text := Text{
-			S: textBuilder.String(),
-			X: currentX,
-			Y: currentY,
-		}
-
-		var currentColumn *Column
-		columnFound := false
-		for _, column := range result {
-			if int64(currentX) == column.Position {
-				currentColumn = column
-				columnFound = true
-				break
-			}
-		}
-
-		if !columnFound {
-			currentColumn = &Column{
-				Position: int64(currentX),
-				Content:  TextVertical{},
-			}
-			result = append(result, currentColumn)
-		}
-
-		currentColumn.Content = append(currentColumn.Content, text)
-	}
-
-	p.walkTextBlocks(showText)
-
-	for _, column := range result {
-		sort.Stable(column.Content)
-	}
-
-	sort.Slice(result, func(i, j int) bool {
-		return result[i].Position < result[j].Position
-	})
-
-	return result, err
-}
-
-// Row represents the contents of a row
-type Row struct {
-	Position int64
-	Content  TextHorizontal
-}
-
-// Rows is a list of rows
-type Rows []*Row
-
-// GetTextByRow returns the page's all text grouped by rows
-func (p Page) GetTextByRow() (Rows, error) {
-	result := Rows{}
-	var err error
-
-	defer func() {
-		if r := recover(); r != nil {
-			result = Rows{}
-			err = errors.New(fmt.Sprint(r))
-		}
-	}()
-
-	showText := func(enc TextEncoding, currentX, currentY float64, s string) {
-		var textBuilder bytes.Buffer
-		for _, ch := range enc.Decode(s) {
-			_, err := textBuilder.WriteRune(ch)
-			if err != nil {
-				panic(err)
-			}
-		}
-
-		// if DebugOn {
-		// 	fmt.Println(textBuilder.String())
-		// }
-
-		text := Text{
-			S: textBuilder.String(),
-			X: currentX,
-			Y: currentY,
-		}
-
-		var currentRow *Row
-		rowFound := false
-		for _, row := range result {
-			if int64(currentY) == row.Position {
-				currentRow = row
-				rowFound = true
-				break
-			}
-		}
-
-		if !rowFound {
-			currentRow = &Row{
-				Position: int64(currentY),
-				Content:  TextHorizontal{},
-			}
-			result = append(result, currentRow)
-		}
-
-		currentRow.Content = append(currentRow.Content, text)
-	}
-
-	p.walkTextBlocks(showText)
-
-	for _, row := range result {
-		sort.Stable(row.Content)
-	}
-
-	sort.Slice(result, func(i, j int) bool {
-		return result[i].Position > result[j].Position
-	})
-
-	return result, err
 }
 
 // walkState holds mutable state for the walkTextBlocks Interpret callback.
@@ -1308,96 +722,4 @@ func (p Page) Content() Content {
 	}
 	Interpret(p.V.Key("Contents"), s.interpret)
 	return Content{s.text, s.rect}
-}
-
-// TextVertical implements sort.Interface for sorting
-// a slice of Text values in vertical order, top to bottom,
-// and then left to right within a line.
-type TextVertical []Text
-
-func (x TextVertical) Len() int      { return len(x) }
-func (x TextVertical) Swap(i, j int) { x[i], x[j] = x[j], x[i] }
-func (x TextVertical) Less(i, j int) bool {
-	if x[i].Y != x[j].Y {
-		return x[i].Y > x[j].Y
-	}
-	return x[i].X < x[j].X
-}
-
-// TextHorizontal implements sort.Interface for sorting
-// a slice of Text values in horizontal order, left to right,
-// and then top to bottom within a column.
-type TextHorizontal []Text
-
-func (x TextHorizontal) Len() int      { return len(x) }
-func (x TextHorizontal) Swap(i, j int) { x[i], x[j] = x[j], x[i] }
-func (x TextHorizontal) Less(i, j int) bool {
-	if x[i].X != x[j].X {
-		return x[i].X < x[j].X
-	}
-	return x[i].Y > x[j].Y
-}
-
-// An Outline is a tree describing the outline (also known as the table of contents)
-// of a document.
-type Outline struct {
-	Title string    // title for this element
-	Page  int       // 1-based page number; 0 if destination cannot be resolved
-	Child []Outline // child elements
-}
-
-// Outline returns the document outline.
-// The Outline returned is the root of the outline tree and typically has no Title itself.
-// That is, the children of the returned root are the top-level entries in the outline.
-func (r *Reader) Outline() Outline {
-	pages := r.buildPageMap()
-	return buildOutline(r.Trailer().Key("Root").Key("Outlines"), pages)
-}
-
-func (r *Reader) buildPageMap() map[uint32]int {
-	m := make(map[uint32]int)
-	n := r.NumPage()
-	for i := 1; i <= n; i++ {
-		p := r.Page(i)
-		if !p.V.IsNull() {
-			m[p.V.ptr.id] = i
-		}
-	}
-	return m
-}
-
-func buildOutline(entry Value, pages map[uint32]int) Outline {
-	var x Outline
-	x.Title = entry.Key("Title").Text()
-	x.Page = resolveOutlineDest(entry, pages)
-	for child := entry.Key("First"); child.Kind() == Dict; child = child.Key("Next") {
-		x.Child = append(x.Child, buildOutline(child, pages))
-	}
-	return x
-}
-
-func resolveOutlineDest(entry Value, pages map[uint32]int) int {
-	dest := entry.Key("Dest")
-	if dest.Kind() == Array {
-		return pageFromDestArray(dest, pages)
-	}
-	action := entry.Key("A")
-	if action.Key("S").Name() == "GoTo" {
-		d := action.Key("D")
-		if d.Kind() == Array {
-			return pageFromDestArray(d, pages)
-		}
-	}
-	return 0
-}
-
-func pageFromDestArray(dest Value, pages map[uint32]int) int {
-	if dest.Len() == 0 {
-		return 0
-	}
-	pageVal := dest.Index(0)
-	if n, ok := pages[pageVal.ptr.id]; ok {
-		return n
-	}
-	return 0
 }
